requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
supabase==1.1.1
psycopg2-binary==2.9.9
python-dateutil==2.8.2
PyJWT==2.8.0
requests==2.31.0